backend/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
faiss-cpu==1.11.0.post1
sentence-transformers==5.1.0
numpy==1.26.4
groq==1.0.0
python-dotenv==1.0.0
pydantic==2.11.3
